world matrix auto updates off for scene
