+++ packages/engine/src/scene/classes/InfiniteGridHelper.ts
@@ -65,8 +65,6 @@ export default class InfiniteGridHelper extends Mesh {
   plane: Plane
   intersectionPointWorld: Vector3
   intersection: any
-  matrixAutoUpdate = false
-  matrixWorldAutoUpdate = false
 
   constructor(size1?, size2?, color?, distance?) {
     color = color || new Color('white')
